make it compile with gcc 7
With gcc 7 more checks will be introduced, among others a:
  __builtin___snprintf_chk
which checks the snprintf-parameters. This check results in:
 alljoyn_core/samples/secure/DeskTopSharedKSService.cc:161:10: error: '__builtin___snprintf_chk' output may be truncated before the last format character [-Werror=format-truncation=]
      bool RequestCredentials(const char* authMechanism, const char* authPeer, uint16_t authCount, const char* userId, uint16_t credMask, Credentials& creds) {
           ^~~~~~~~~~~~~~~~~~
 In file included from /usr/include/stdio.h:938:0,
                  from alljoyn_core/samples/secure/DeskTopSharedKSService.cc:28:
 /usr/include/x86_64-linux-gnu/bits/stdio2.h:65:44: note: '__builtin___snprintf_chk' output between 7 and 8 bytes into a destination of size 7
         __bos (__s), __fmt, __va_arg_pack ());

Though rand() does only returns values between 0 and MAX, it is handled
as 'int' and though the check assumes that 7 chars of pinStr are not
enough to handle all possible values.
Casting pin to 'unsigned int' solves this problem.

+++ alljoyn_core/samples/secure/DeskTopSharedKSService.cc
@@ -192,14 +192,14 @@ class SrpKeyXListener : public AuthListener {
                 if (authCount <= 3) {
                     /* seed the random number */
                     srand(time(nullptr));
-                    int pin = rand() % 1000000;
+                    unsigned int pin = (unsigned int)rand() % 1000000;
                     char pinStr[7];
 
 #if defined(QCC_OS_GROUP_WINDOWS)
-                    _snprintf(pinStr, ArraySize(pinStr), "%06d", pin);
+                    _snprintf(pinStr, ArraySize(pinStr), "%06u", pin);
                     pinStr[ArraySize(pinStr) - 1] = '\0';
 #else
-                    snprintf(pinStr, ArraySize(pinStr), "%06d", pin);
+                    snprintf(pinStr, ArraySize(pinStr), "%06u", pin);
 #endif
 
                     printf("One Time Password : %s\n", pinStr);
